three syntactic categories
Conor implements a dependently-typed language, so types are terms.
I don't want to be distracted by the peculiarities of dependent types, I
want to focus on the worlds, so I'll start with the plain lambda
calculus, plus the unit type so function have something to manipulate.

Like Conor, I'll divide my terms into two syntactic categories, but I
don't think "constructions" and "eliminations" are the right names for
them. After all, why would we check that UnitV has type UnitT? We can
synthesize this type easily. Instead, let's put all the terms in the
synthesized category if we can, so we can infer as much as possible, and
let's put all the sub-terms in the checked category if we can, so we can
use the inferred information and avoid type annotations as much as
possible.

Thus, I have the two syntactic categories above, plus types, for a total
of three syntactic categories.